+++ test/Microsoft.AspNet.Server.KestrelTests/MemoryPoolExtensions.cs
@@ -1,10 +1,10 @@
-﻿using Microsoft.AspNet.Server.Kestrel.Http;
+﻿using Microsoft.AspNet.Server.Kestrel.Infrastructure;
 
 namespace Microsoft.AspNet.Server.KestrelTests
 {
     public static class MemoryPoolExtensions
     {
-        public static MemoryPoolBlock2.Iterator Add(this MemoryPoolBlock2.Iterator iterator, int count)
+        public static MemoryPoolIterator2 Add(this MemoryPoolIterator2 iterator, int count)
         {
             int actual;
             return iterator.CopyTo(new byte[count], 0, count,  out actual);
